requirements.txt
flask
opencv-python
numpy
shapely
Pillow
requests
playwright
